// app-ui/src/app/backend/model/schemaTdv.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SchemaTdv { 
    COLUMN_NAME: string;
    DATA_TYPE: string;
    ORDINAL_POSITION: number;
    JDBC_DATA_TYPE: number;
}